Update namespaces based on feedback
