+++ gems/smithy-client/lib/smithy-client.rb
@@ -45,10 +45,12 @@ require_relative 'smithy-client/structure'
 require_relative 'smithy-client/endpoint_rules'
 
 # codecs
+require_relative 'smithy-client/codec/cbor'
+
+# cbor
 require_relative 'smithy-client/cbor'
 require_relative 'smithy-client/cbor/encoder'
 require_relative 'smithy-client/cbor/decoder'
-require_relative 'smithy-client/cbor_codec'
 
 module Smithy
   # Base module for a generated Smithy gem.
