+++ gems/smithy-client/lib/smithy-client/codec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative 'codec/cbor'
+
+module Smithy
+  module Client
+    # Contains different serialization and deserialization formats
+    module Codec
+    end
+  end
+end
