--- gems/smithy-client/sig/smithy-client/cbor_codec.rbs
@@ -1,12 +1,0 @@
-module Smithy
-  module Client
-    class CBORCodec
-      def initialize: (?::Hash[untyped, untyped] options) -> void
-
-      def serialize: (untyped data, Shapes::Shape shape) -> (nil | String)
-
-      def deserialize: (String bytes, Shapes::Shape shape, ?Struct[untyped] target) -> (Hash[untyped, untyped] | Object)
-      
-    end
-  end
-end
